% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coindeskr.R
\name{get_currency_list}
\alias{get_currency_list}
\title{Extract List of currency codes supported by Coindesk}
\usage{
get_currency_list()
}
\value{
Dataframe of different currency codes supported by Coindesk along with its country
}
\description{
Extract List of currency codes supported by Coindesk
}
\examples{
get_currency_list()
}
