% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coindeskr.R
\name{get_last31days_price}
\alias{get_last31days_price}
\title{Extract daily USD Price of Bitcoin for the last 31 days}
\usage{
get_last31days_price()
}
\value{
Dataframe with USD Price as one column and Date as rownames
}
\description{
Extract daily USD Price of Bitcoin for the last 31 days
}
\examples{
get_last31days_price()
}
