% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coindeskr.R
\name{get_current_price}
\alias{get_current_price}
\title{Extract Bitcoin Price Index from Coindesk API}
\usage{
get_current_price()
}
\value{
structured list of Bitcoin price in different currencies returned from the API call
}
\description{
Extract Bitcoin Price Index from Coindesk API
}
\examples{
get_current_price()
}
